+++ src/config/mysql-setting.js
@@ -3,7 +3,6 @@ const fs = require('fs');
 const setting = {
   db_name: {
     dev: process.env.SQL_DATABASE
-    // dev: 'Page'
   },
   db_user: process.env.SQL_USER,
   db_password: process.env.SQL_PASSWORD,
@@ -22,9 +21,9 @@ if (process.env.INSTANCE_CONNECTION_NAME && process.env.NODE_ENV === 'production
 } else {
   setting.host = '35.188.137.1';
   const SSL = {
-    key: fs.readFileSync(path.join(__dirname, '../auth/client-key.pem')),
-    cert: fs.readFileSync(path.join(__dirname, '../auth/client-cert.pem')),
-    ca: fs.readFileSync(path.join(__dirname, '../auth/server-ca.pem')),
+    key: fs.readFileSync(path.join(__dirname, process.env.SQL_SSL_KEY_PATH)),
+    cert: fs.readFileSync(path.join(__dirname, process.env.SQL_SSL_CERT_PATH)),
+    ca: fs.readFileSync(path.join(__dirname, process.env.SQL_SSL_CA_PATH)),
     // ref: https://stackoverflow.com/questions/52465530/sequelize-connection-timeout-while-using-serverless-aurora-looking-for-a-way-to
     connectTimeout: 60000 // The milliseconds before a timeout occurs during the initial connection to the MySQL server.
   };