+++ src/model/livescore/prematchBaseballModel.js
@@ -3,6 +3,7 @@ const firebaseAdmin = require('../../util/firebaseUtil');
 const firestore = firebaseAdmin().firestore();
 const db = require('../../util/dbUtil');
 const AppErrors = require('../../util/AppErrors');
+const { logger } = require('firebase-functions');
 const limit = 10;
 
 function prematchBaseball(args) {
@@ -25,13 +26,13 @@ function getHomeAndAwayTeamFromMySQL(args) {
     try {
       const result = await db.sequelize.query(
         // index is const, except match__seasons, match__spreads, match__totals table is ref, taking 170ms
-        `SELECT game.bets_id, game.league_id, game.home_id, game.away_id, game.season, game.status, game.scheduled,
-                home.name AS home_name, home.name_ch AS home_name_ch, home.alias AS home_alias, home.alias_ch AS home_alias_ch, home.image_id AS home_image_id, 
-                away.name AS away_name, away.name_ch AS away_name_ch, away.alias AS away_alias, away.alias_ch AS away_alias_ch, away.image_id AS away_image_id, 
+        `SELECT game.bets_id, game.league_id, game.home_id, game.away_id, game.season, game.status, game.scheduled, game.home_player, game.away_player, 
+                home.name AS home_name, home.name_ch AS home_name_ch, home.alias AS home_alias, home.alias_ch AS home_alias_ch, home.image_id AS home_image_id, home.team_id AS home_team_id, 
+                away.name AS away_name, away.name_ch AS away_name_ch, away.alias AS away_alias, away.alias_ch AS away_alias_ch, away.image_id AS away_image_id, away.team_id AS away_team_id, 
                 spread.spread_id, spread.handicap AS spread_handicap, spread.home_tw, spread.away_tw, spread.rate AS spread_rate, 
                 totals.totals_id, totals.handicap AS totals_handicap, totals.over_tw, totals.rate AS totals_rate
           FROM (
-                  SELECT matches.bets_id, matches.league_id, matches.spread_id, matches.totals_id, matches.home_id, matches.away_id, matches.status, matches.scheduled, season.season
+                  SELECT matches.bets_id, matches.league_id, matches.spread_id, matches.totals_id, matches.home_id, matches.away_id, matches.status, matches.scheduled, matches.home_player, matches.away_player, season.season
                     FROM matches
                LEFT JOIN match__seasons AS season ON season.league_id = matches.league_id
                    WHERE matches.scheduled BETWEEN UNIX_TIMESTAMP(season.start_date) AND UNIX_TIMESTAMP(season.end_date)
@@ -53,7 +54,7 @@ function getHomeAndAwayTeamFromMySQL(args) {
 
       !result.length ? reject(new AppErrors.MatchNotFound()) : resolve(result[0]);
     } catch (err) {
-      return reject(new AppErrors.MysqlError(`${err.stack} by TsaiChieh`));
+      return reject(new AppErrors.MysqlError(err.stack));
     }
   });
 }
@@ -68,7 +69,7 @@ function getPrematchFromFirestore(args, matchData) {
       if (!homeData.exists || !awayData.exists) return reject(new AppErrors.TeamInformationNotFound());
       return resolve({ homeData: homeData.data()[`season_${season}`], awayData: awayData.data()[`season_${season}`] });
     } catch (err) {
-      return reject(new AppErrors.FirestoreQueryError(`${err.stack} by TsaiChieh`));
+      return reject(new AppErrors.FirestoreQueryError(err.stack));
     }
   });
 }
@@ -80,7 +81,7 @@ function queryHomeEvents(args) {
         // take 169 ms
         `(
           SELECT game.home_id AS aim_home_id, game.away_id AS aim_away_id,
-                 historygame.bets_id AS id, historygame.scheduled AS scheduled,  historygame.home_id AS history_home_id, historygame.away_id AS history_away_id,
+                 historygame.bets_id AS id, historygame.scheduled AS scheduled, historygame.home_id AS history_home_id, historygame.away_id AS history_away_id,
                  historygame.spread_result AS history_spread_result, historygame.totals_result AS history_totals_result, spread.handicap AS spread_handicap, totals.handicap AS totals_handicap
             FROM matches AS game,
                  matches AS historygame,
@@ -97,7 +98,7 @@ function queryHomeEvents(args) {
              AND historygame.spread_id = spread.spread_id
              AND historygame.bets_id = totals.match_id
              AND historygame.totals_id = totals.totals_id
-        ORDER BY historygame.scheduled      
+        ORDER BY historygame.scheduled DESC     
         LIMIT ${limit}
         )`,
         {
@@ -109,9 +110,9 @@ function queryHomeEvents(args) {
         }
       );
 
-      return resolve(await queries);
+      return resolve(queries);
     } catch (err) {
-      return reject(`${err.stack} by DY`);
+      return reject(new AppErrors.MysqlError(err.stack));
     }
   });
 }
@@ -140,7 +141,7 @@ function queryAwayEvents(args) {
            AND historygame.spread_id = spread.spread_id
            AND historygame.bets_id = totals.match_id
            AND historygame.totals_id = totals.totals_id
-      ORDER BY historygame.scheduled
+      ORDER BY historygame.scheduled DESC
          LIMIT ${limit}        
         )`,
         {
@@ -152,9 +153,9 @@ function queryAwayEvents(args) {
         }
       );
 
-      return resolve(await queries);
+      return resolve(queries);
     } catch (err) {
-      return reject(`${err.stack} by DY`);
+      return reject(new AppErrors.MysqlError(err.stack));
     }
   });
 }
@@ -163,10 +164,7 @@ function repackagePrematch(args, teamsFromFirestore, teamsFromMySQL, events, fig
   const { homeData, awayData } = teamsFromFirestore;
   const rate = repackagePassRate(events);
   const tenFights = repackageTenFights(args, fights);
-
-  // FIXME 目前針對 MLB 聯盟作 null 判斷
-  const homeDataIsNull = homeData === null;
-  const awayDataIsNull = awayData === null;
+  const { homePlayer, awayPlayer, homePlayerIsNull, awayPlayerIsNull } = checkPlayer(teamsFromMySQL);
 
   try {
     const data = {
@@ -195,44 +193,38 @@ function repackagePrematch(args, teamsFromFirestore, teamsFromMySQL, events, fig
         alias: teamsFromMySQL.home_alias,
         alias_ch: teamsFromMySQL.home_alias_ch,
         team_name: teamsFromMySQL.home_alias_ch,
+        team_id: teamsFromMySQL.home_team_id,
         name: teamsFromMySQL.home_name,
         name_ch: teamsFromMySQL.home_name_ch,
         image_id: teamsFromMySQL.home_image_id,
         team_base: {
-          spread_rate: rate.home_spread_rate,
-          totals_rate: rate.home_totals_rate,
-          L10: !homeDataIsNull ? homeData.team_base.L10 : null,
-          STRK: !homeDataIsNull ? homeData.team_base.STRK : null,
-          Win: !homeDataIsNull ? homeData.team_base.Win : null,
-          Loss: !homeDataIsNull ? homeData.team_base.Loss : null,
-          Draw: !homeDataIsNull ? homeData.team_base.Draw : null,
-          at_home: !homeDataIsNull ? homeData.team_base.at_home : null,
-          at_away: !homeDataIsNull ? homeData.team_base.at_away : null,
-          per_R: !homeDataIsNull ? homeData.team_base.per_R : null,
-          per_allow_R: !homeDataIsNull ? homeData.team_base.per_allow_R : null
-          // L10: homeData.team_base.L10,
-          // STRK: homeData.team_base.STRK,
-          // Win: homeData.team_base.Win,
-          // Loss: homeData.team_base.Loss,
-          // Draw: homeData.team_base.Draw,
-          // at_home: homeData.team_base.at_home,
-          // at_away: homeData.team_base.at_away,
-          // per_R: homeData.team_base.per_R,
-          // allow_per_R: homeData.team_base.per_allow_R
+          spread_rate: Number((rate.home_spread_rate).toFixed(2)),
+          totals_rate: Number((rate.home_totals_rate).toFixed(2)),
+          L10: homeData.team_base.L10,
+          STRK: homeData.team_base.STRK,
+          Win: homeData.team_base.Win,
+          Loss: homeData.team_base.Loss,
+          Draw: homeData.team_base.Draw,
+          at_home: homeData.team_base.at_home,
+          at_away: homeData.team_base.at_away,
+          per_R: homeData.team_base.per_R,
+          allow_per_R: homeData.team_base.per_allow_R
+        },
+        pitcher: {
+          id: homePlayerIsNull ? null : homePlayer.pitchers.id,
+          Win: homePlayerIsNull ? null : homePlayer.pitchers.Win,
+          Loss: homePlayerIsNull ? null : homePlayer.pitchers.Loss,
+          ERA: homePlayerIsNull ? null : homePlayer.pitchers.ERA,
+          SO: homePlayerIsNull ? null : homePlayer.pitchers.SO,
+          jersey_id: homePlayerIsNull ? null : homePlayer.pitchers.jersey_id
         },
         team_hit: {
-          R: !homeDataIsNull ? homeData.team_hit.R : null,
-          H: !homeDataIsNull ? homeData.team_hit.H : null,
-          HR: !homeDataIsNull ? homeData.team_hit.HR : null,
-          AVG: !homeDataIsNull ? homeData.team_hit.AVG : null,
-          OBP: !homeDataIsNull ? homeData.team_hit.OBP : null,
-          SLG: !homeDataIsNull ? homeData.team_hit.SLG : null
-          // R: homeData.team_hit.R,
-          // H: homeData.team_hit.H,
-          // HR: homeData.team_hit.HR,
-          // AVG: homeData.team_hit.AVG,
-          // OBP: homeData.team_hit.OBP,
-          // SLG: homeData.team_hit.SLG
+          R: homeData.team_hit.R,
+          H: homeData.team_hit.H,
+          HR: homeData.team_hit.HR,
+          AVG: homeData.team_hit.AVG,
+          OBP: homeData.team_hit.OBP,
+          SLG: homeData.team_hit.SLG
         }
       },
       away: {
@@ -240,54 +232,67 @@ function repackagePrematch(args, teamsFromFirestore, teamsFromMySQL, events, fig
         alias: teamsFromMySQL.away_alias,
         alias_ch: teamsFromMySQL.away_alias_ch,
         team_name: teamsFromMySQL.away_alias_ch,
+        team_id: teamsFromMySQL.away_team_id,
         name: teamsFromMySQL.away_name,
         name_ch: teamsFromMySQL.away_name_ch,
         image_id: teamsFromMySQL.away_image_id,
         team_base: {
-          spread_rate: rate.away_spread_rate,
-          totals_rate: rate.away_totals_rate,
-          L10: !awayDataIsNull ? awayData.team_base.L10 : null,
-          STRK: !awayDataIsNull ? awayData.team_base.STRK : null,
-          Win: !awayDataIsNull ? awayData.team_base.Win : null,
-          Loss: !awayDataIsNull ? awayData.team_base.Loss : null,
-          Draw: !awayDataIsNull ? awayData.team_base.Draw : null,
-          at_home: !awayDataIsNull ? awayData.team_base.at_home : null,
-          at_away: !awayDataIsNull ? awayData.team_base.at_away : null,
-          per_R: !awayDataIsNull ? awayData.team_base.per_R : null,
-          per_allow_R: !awayDataIsNull ? awayData.team_base.per_allow_R : null
-          // L10: awayData.team_base.L10,
-          // STRK: awayData.team_base.STRK,
-          // Win: awayData.team_base.Win,
-          // Loss: awayData.team_base.Loss,
-          // Draw: awayData.team_base.Draw,
-          // at_home: awayData.team_base.at_home,
-          // at_away: awayData.team_base.at_away,
-          // per_R: awayData.team_base.per_R,
-          // allow_per_R: awayData.team_base.allow_per_R
+          spread_rate: Number((rate.away_spread_rate).toFixed(2)),
+          totals_rate: Number((rate.away_totals_rate).toFixed(2)),
+          L10: awayData.team_base.L10,
+          STRK: awayData.team_base.STRK,
+          Win: awayData.team_base.Win,
+          Loss: awayData.team_base.Loss,
+          Draw: awayData.team_base.Draw,
+          at_home: awayData.team_base.at_home,
+          at_away: awayData.team_base.at_away,
+          per_R: awayData.team_base.per_R,
+          allow_per_R: awayData.team_base.allow_per_R
+        },
+        pitcher: {
+          id: awayPlayerIsNull ? null : awayPlayer.pitchers.id,
+          Win: awayPlayerIsNull ? null : awayPlayer.pitchers.Win,
+          Loss: awayPlayerIsNull ? null : awayPlayer.pitchers.Loss,
+          ERA: awayPlayerIsNull ? null : awayPlayer.pitchers.ERA,
+          SO: awayPlayerIsNull ? null : awayPlayer.pitchers.SO,
+          jersey_id: awayPlayerIsNull ? null : awayPlayer.pitchers.jersey_id
         },
         team_hit: {
-          R: !awayDataIsNull ? awayData.team_hit.R : null,
-          H: !awayDataIsNull ? awayData.team_hit.H : null,
-          HR: !awayDataIsNull ? awayData.team_hit.HR : null,
-          AVG: !awayDataIsNull ? awayData.team_hit.AVG : null,
-          OBP: !awayDataIsNull ? awayData.team_hit.OBP : null,
-          SLG: !awayDataIsNull ? awayData.team_hit.SLG : null
-          // R: awayData.team_hit.R,
-          // H: awayData.team_hit.H,
-          // HR: awayData.team_hit.HR,
-          // AVG: awayData.team_hit.AVG,
-          // OBP: awayData.team_hit.OBP,
-          // SLG: awayData.team_hit.SLG
+          R: awayData.team_hit.R,
+          H: awayData.team_hit.H,
+          HR: awayData.team_hit.HR,
+          AVG: awayData.team_hit.AVG,
+          OBP: awayData.team_hit.OBP,
+          SLG: awayData.team_hit.SLG
         }
       },
       L10_record: tenFights
-
     };
     return data;
   } catch (err) {
-    console.error(err);
-    throw new AppErrors.RepackageError(`${err.stack} by TsaiChieh`);
+    throw new AppErrors.RepackageError(err.stack);
   }
+}
+
+function checkPlayer(teamsFromMySQL) {
+  const homePlayer = JSON.parse(teamsFromMySQL.home_player);
+  const awayPlayer = JSON.parse(teamsFromMySQL.away_player);
+  let homePlayerIsNull = true;
+  let awayPlayerIsNull = true;
+
+  if (homePlayer) {
+    homePlayerIsNull = homePlayer.pitchers === null;
+    if (homePlayer.pitchers) {
+      if (homePlayer.pitchers.id === 0) homePlayerIsNull = true;
+    }
+  }
+  if (awayPlayer) {
+    awayPlayerIsNull = awayPlayer.pitchers === null;
+    if (awayPlayer.pitchers) {
+      if (awayPlayer.pitchers.id === 0) awayPlayerIsNull = true;
+    }
+  }
+  return { homePlayer, awayPlayer, homePlayerIsNull, awayPlayerIsNull };
 }
 
 function repackagePassRate(events) {
@@ -498,10 +503,10 @@ function repackagePassRate(events) {
       }
     }
     return {
-      home_spread_rate: homeAtGivePass / 10,
-      home_totals_rate: homeAtOverPass / 10,
-      away_spread_rate: awayAtGivePass / 10,
-      away_totals_rate: awayAtOverPass / 10
+      home_spread_rate: (homeAtGivePass + homeAtBeGivenPass) / homeEvents.length,
+      home_totals_rate: homeAtOverPass / homeEvents.length,
+      away_spread_rate: (awayAtGivePass + awayAtBeGivenPass) / awayEvents.length,
+      away_totals_rate: awayAtOverPass / awayEvents.length
     };
   } catch (err) {
     console.error(err);
@@ -599,9 +604,9 @@ function queryTenFightEvent(args) {
           type: db.sequelize.QueryTypes.SELECT
         }
       );
-      return resolve(await queries);
+      return resolve(queries);
     } catch (err) {
-      return reject(new AppErrors.MysqlError(`${err.stack} by DY`));
+      return reject(new AppErrors.MysqlError(err.stack));
     }
   });
 }
@@ -692,8 +697,8 @@ function repackageTenFights(args, events) {
 
     return data;
   } catch (err) {
-    console.error(`${err.stack} by DY`);
-    throw new AppErrors.RepackageError(`${err.stack} by DY`);
+    logger.error(err.stack);
+    throw new AppErrors.RepackageError(err.stack);
   }
 }