+++ src/routers/user.js
@@ -222,7 +222,7 @@ router.post(
 /* 大神結算 */
 router.post(
   '/settle_god_rank',
-  verification.token,
+  verification.token_v2,
   require('../controller/user/settleGodRankController')
 );